requirements.txt
asgiref==3.4.0
certifi==2021.5.30
chardet==4.0.0
dj-database-url==0.5.0
Django==3.2.4
django-cors-headers==3.7.0
django-heroku==0.3.1
django-rest-framework==0.1.0
djangorestframework==3.12.4
gunicorn==20.1.0
idna==2.10
joblib==1.0.1
numpy==1.21.0
pandas==1.3.0
psycopg2==2.9.1
psycopg2-binary==2.9.1
python-dateutil==2.8.2
python-dotenv==0.19.0
pytz==2021.1
requests==2.25.1
scikit-learn==0.24.2
scipy==1.7.0
six==1.16.0
sklearn==0.0
sqlparse==0.4.1
threadpoolctl==2.2.0
urllib3==1.26.6
whitenoise==5.3.0
